+++ easymanager/easymanager-core/src/test/java/progettotlp/persistenza/DdTManagerImplFunctionalTest.java
@@ -6,8 +6,6 @@
 package progettotlp.persistenza;
 
 import java.util.ArrayList;
-import org.junit.After;
-import org.junit.Before;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
@@ -40,12 +38,12 @@ public class DdTManagerImplFunctionalTest extends AbstractTest{
         b1.setCodice("cod");
         b1.setCommessa("com");
         b1.setDescrizione("descr");
-        b1.setQta(1);
+        b1.setQta(1F);
         Bene b2 = new Bene();
         b2.setCodice("cod2");
         b2.setCommessa("com2");
         b2.setDescrizione("descr2");
-        b2.setQta(2);
+        b2.setQta(2F);
         DdT toSave = new DdT();
         toSave.setBeni(Arrays.asList(b1, b2));
         Date data = DateUtils.parseDate("01-01-2012");
@@ -66,12 +64,12 @@ public class DdTManagerImplFunctionalTest extends AbstractTest{
         assertEquals("cod", b1.getCodice());
         assertEquals("com", b1.getCommessa());
         assertEquals("descr", b1.getDescrizione());
-        assertEquals(new Integer(1), b1.getQta());
+        assertEquals(new Float(1), b1.getQta());
         b2 = beni.get(1);
         assertEquals("cod2", b2.getCodice());
         assertEquals("com2", b2.getCommessa());
         assertEquals("descr2", b2.getDescrizione());
-        assertEquals(new Integer(2), b2.getQta());
+        assertEquals(new Float(2), b2.getQta());
         assertEquals(data, retrieved.getData());
         assertEquals(cliente, retrieved.getCliente());
         assertEquals("Cessionario", retrieved.getMezzo());
@@ -91,7 +89,7 @@ public class DdTManagerImplFunctionalTest extends AbstractTest{
             Bene b=beni.get(i);
             Long id = b.getId();
             if (id.equals(1L)) {
-                b.setQta(1);
+                b.setQta(1F);
                 b.setCampionario(Boolean.TRUE);
                 b.setPiazzato(Boolean.TRUE);
                 b.setPrimoCapo(Boolean.TRUE);
@@ -102,7 +100,7 @@ public class DdTManagerImplFunctionalTest extends AbstractTest{
                 fail();
             }
         }
-        Bene beneToAdd = new Bene("cod3", "com3", "descr3", 30, true, true, true, true,true);
+        Bene beneToAdd = new Bene("cod3", "com3", "descr3", 30F, true, true, true, true,true);
         beni.add(beneToAdd);
 
         ddTManager.modificaDdT(toModify);
@@ -112,7 +110,7 @@ public class DdTManagerImplFunctionalTest extends AbstractTest{
         assertEquals(2, beni.size());
         for (Bene b : beni) {
             if (b.getId().equals(1L)) {
-                assertEquals(new Integer(1), b.getQta());
+                assertEquals(new Float(1), b.getQta());
                 assertEquals(new Float(21), b.getPrezzo());
                 assertTrue(b.getCampionario());
                 assertTrue(b.getPiazzato());
